[TIP] Reorganize barchart folder within indicators module
