+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/barchart/barchart/barchart.tsx
@@ -9,9 +9,9 @@ import React, { VFC } from 'react';
 import { Axis, BarSeries, Chart, Position, ScaleType, Settings } from '@elastic/charts';
 import { EuiThemeProvider } from '@elastic/eui';
 import { TimeRangeBounds } from '@kbn/data-plugin/common';
-import { IndicatorBarchartLegendAction } from '../indicator_barchart_legend_action/indicator_barchart_legend_action';
-import { barChartTimeAxisLabelFormatter } from '../../../../common/utils/dates';
-import type { ChartSeries } from '../../services/fetch_aggregated_indicators';
+import { IndicatorBarchartLegendAction } from '../legend_action';
+import { barChartTimeAxisLabelFormatter } from '../../../../../common/utils/dates';
+import type { ChartSeries } from '../../../services';
 
 const ID = 'tiIndicator';
 const DEFAULT_CHART_HEIGHT = '200px';
