+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/barchart/field_selector/styles.ts
