+++ x-pack/plugins/threat_intelligence/public/modules/indicators/hooks/index.ts
@@ -5,4 +5,7 @@
  * 2.0.
  */
 
-export * from './indicators_field_selector';
+export * from './use_aggregated_indicators';
+export * from './use_indicators_filters_context';
+export * from './use_indicators_total_count';
+export * from './use_sourcerer_data_view';
