+++ x-pack/plugins/threat_intelligence/public/modules/indicators/hooks/use_aggregated_indicators.test.tsx
@@ -12,7 +12,7 @@ import {
   mockedTimefilterService,
   TestProvidersComponent,
 } from '../../../common/mocks/test_providers';
-import { createFetchAggregatedIndicators } from '../services/fetch_aggregated_indicators';
+import { createFetchAggregatedIndicators } from '../services';
 
 jest.mock('../services/fetch_aggregated_indicators');
